I'm at the 6-hour limit but didn't finish.
Reserve and Cancel buttons do not work. No chance to deploy or do the readme. But the backend functions are all complete and working. And Database Tests are complete and working.

Not long to go, but I think the 6-hour time limit may be small on purpose. I did hustle, just not everything works perfectly the very first time you try.